# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Café Lauren",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/queue": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Get Queue",
        "operationId": "getQueue",
        "parameters": [
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QueueItem"
                  },
                  "title": "Response Getqueue"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Add To Queue",
        "operationId": "addToQueue",
        "parameters": [
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueAddRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QueueItem"
                  },
                  "title": "Response Addtoqueue"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/queue/{recipe_id}": {
      "delete": {
        "tags": [
          "queue"
        ],
        "summary": "Remove From Queue",
        "operationId": "removeFromQueue",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QueueItem"
                  },
                  "title": "Response Removefromqueue"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requests": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "List Requests",
        "operationId": "listRequests",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "new",
                "planned",
                "declined",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestOut"
                  },
                  "title": "Response Listrequests"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Create Request",
        "operationId": "createRequest",
        "parameters": [
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requests/{request_id}/answer": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Answer Request",
        "operationId": "answerRequest",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAnswer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/staples": {
      "get": {
        "tags": [
          "staples"
        ],
        "summary": "List Staples",
        "operationId": "listStaples",
        "parameters": [
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StapleOut"
                  },
                  "title": "Response Liststaples"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "staples"
        ],
        "summary": "Create Staple",
        "operationId": "createStaple",
        "parameters": [
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StapleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StapleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/staples/{staple_id}": {
      "patch": {
        "tags": [
          "staples"
        ],
        "summary": "Patch Staple",
        "operationId": "patchStaple",
        "parameters": [
          {
            "name": "staple_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Staple Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaplePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StapleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "staples"
        ],
        "summary": "Delete Staple",
        "operationId": "deleteStaple",
        "parameters": [
          {
            "name": "staple_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Staple Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "Ok"
      },
...
      "QueueAddRequest": {
        "properties": {
          "recipe_id": {
            "type": "integer",
            "title": "Recipe Id"
          }
        },
        "type": "object",
        "required": [
          "recipe_id"
        ],
        "title": "QueueAddRequest"
      },
      "QueueItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "recipe": {
            "$ref": "#/components/schemas/RecipeSummary"
          },
          "by": {
            "type": "string",
            "enum": [
              "lauren",
              "joe",
              "leidy"
            ],
            "title": "By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recipe",
          "by",
          "created_at"
        ],
        "title": "QueueItem"
      },
...
      "RecipeSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "short_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "total_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Min"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "healthy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthy"
          },
          "delicious": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delicious"
          },
          "stars": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stars"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source": {
            "type": "string",
            "enum": [
              "imported",
              "ai",
              "manual"
            ],
            "title": "Source"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "saved"
            ],
            "title": "Status"
          },
          "last_made": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Made"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source",
          "status"
        ],
        "title": "RecipeSummary"
      },
...
      "RequestAnswer": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "new",
              "planned",
              "declined"
            ],
            "title": "Status"
          },
          "reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply"
          },
          "list_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "List Items",
            "description": "Free-text items (\"2 lbs apples\") to add to this week's list."
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "RequestAnswer"
      },
...
      "RequestCreate": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "meal",
              "out"
            ],
            "title": "Type"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "type",
          "text"
        ],
        "title": "RequestCreate"
      },
      "RequestOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "who": {
            "type": "string",
            "enum": [
              "lauren",
              "joe",
              "leidy"
            ],
            "title": "Who"
          },
          "type": {
            "type": "string",
            "enum": [
              "meal",
              "out"
            ],
            "title": "Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "planned",
              "declined"
            ],
            "title": "Status"
          },
          "reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply"
          },
          "week_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "who",
          "type",
          "text",
          "status",
          "created_at"
        ],
        "title": "RequestOut"
      },
...
      "StapleCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "section": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "produce",
                  "frozen",
                  "meat",
                  "dry",
                  "dairy",
                  "beverages"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "StapleCreate"
      },
      "StapleOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "section": {
            "type": "string",
            "enum": [
              "produce",
              "frozen",
              "meat",
              "dry",
              "dairy",
              "beverages"
            ],
            "title": "Section"
          },
          "from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "section",
          "active"
        ],
        "title": "StapleOut"
      },
      "StaplePatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "section": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "produce",
                  "frozen",
                  "meat",
                  "dry",
                  "dairy",
                  "beverages"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "type": "object",
        "title": "StaplePatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}